Changed execution of sudo to be dependent on it's existence
This change closes #3 'need to check if sudo is installed'.

Without some constraint if sudo is not found at '/usr/bin/sudo' there would be
errors:

   error: execresult '/usr/bin/sudo -l -U gnats' is assumed to be executable but isn't
   error: Proposed executable file '/usr/bin/sudo' doesn't exist

This change tries to find the path to sudo by checking for it's definition in
default:paths.sudo before falling back to a reasonable default.

Additionally, we emit a warning if the agent has inform, verbose, or debug
mode classes defined.